requirements.txt
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0
pandas>=1.5.0
openpyxl>=3.0.0
flask>=2.0.0
flask-cors>=3.0.0
beautifulsoup4>=4.10.0